src/requirements.txt
requests
ujson
snowflake-connector-python
